Added in UI Stuff
Added in Health Bar and Armor Bar for player. Also added in the starting cards into Game Controller, when then is transferred to the deck when the method getStartingCards method is called. These needs to be called when player start() method is called, however these dont match up to game objects  are causing errors.